Change length of dif
I noticed here that if we set dif to be length of domains, then it may be that it fails if the length of bcs > length of domains. I don't know if this is intended, but I thought it was wiser to set this to the length of the iterator to avoid errors.