Refactor internals so debugging is not sensitive to the internals
There's still a whole bunch more to do, but this is a major step. The purpose of `symbolic_discretize` is that it's supposed to be the sole entry point for getting the "behind the scenes" of the problem being generated. Clearly that failed to occur in this repo's implementation, with an exhaustive "debugging" interface, which really should just be `symbolic_discretize` with sufficient documentation on its return. The issue also manifested in the fact that huge loads of arguments would get passed around to every function instead of just putting them in a type. This also meant that many functions were repeated, some were never callable, etc.

This fixes a large part of that, with more to be fixed in an upcoming PR. `discretize_inner_functions` is not part of the public API for example so it should show up in 0 documentation lines, but it does so that should be fixed. But since this is a huge step, I'm going to get this completed, tested, merged, and then move onto the next step.