Use dependent variable naming in the indexing of the results
Changes from `depvar_1` for the ith neural network to `x` if it's `[x(t),y(t),z(t)]`. This should make it a lot easier to understand the indexing structure: `res.u.depvar.y` etc. 

This builds on https://github.com/SciML/NeuralPDE.jl/pull/565 and should be the last change for the v5 release.